fix(HTD-44): make tools prop optional and add default value
move example to integrations page
