+++ src/components/sections/integration/single-page/compatibility/compatibility.types.ts
@@ -1,5 +1,5 @@
 export interface IntegrationCompatibilityProps {
   subtitle: string;
   title: string;
-  tools: string[];
+  tools?: string[];
 }
